# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Patna Realty Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leases:
    get:
      operationId: listLeases
      tags: [leases]
      summary: List leases
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: propertyId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [active, expired, terminated]
      responses:
        "200":
          description: List of leases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lease"
    post:
      operationId: createLease
      tags: [leases]
      summary: Create a lease
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeaseBody"
      responses:
        "201":
          description: Lease created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lease"
  /leases/{leaseId}:
    get:
      operationId: getLease
      tags: [leases]
      summary: Get lease by ID
      parameters:
        - name: leaseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lease details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lease"
    put:
      operationId: updateLease
      tags: [leases]
      summary: Update lease
      parameters:
        - name: leaseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeaseBody"
      responses:
        "200":
          description: Updated lease
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lease"
...
  /maintenance:
    get:
      operationId: listMaintenanceRequests
      tags: [maintenance]
      summary: List maintenance requests
      parameters:
        - name: propertyId
          in: query
          schema:
            type: integer
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [open, in_progress, resolved, closed]
      responses:
        "200":
          description: List of requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MaintenanceRequest"
    post:
      operationId: createMaintenanceRequest
      tags: [maintenance]
      summary: Submit maintenance request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMaintenanceBody"
      responses:
        "201":
          description: Request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceRequest"
  /maintenance/{requestId}:
    get:
      operationId: getMaintenanceRequest
      tags: [maintenance]
      summary: Get maintenance request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceRequest"
    put:
      operationId: updateMaintenanceRequest
      tags: [maintenance]
      summary: Update maintenance request status
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMaintenanceBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceRequest"
...
components:
  schemas:
...
    Property:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [flat, shop, grocery_store, market]
        bhk:
          type: ["integer", "null"]
          description: BHK count, applicable only for flats
        sizeSqft:
          type: number
        floor:
          type: ["integer", "null"]
        blockOrSector:
          type: ["string", "null"]
        address:
          type: string
        locality:
          type: string
        landmark:
          type: ["string", "null"]
        rentPrice:
          type: ["number", "null"]
        salePriceINR:
          type: ["number", "null"]
        rentPeriod:
          type: ["string", "null"]
          enum: [daily, weekly, monthly, yearly, null]
        purpose:
          type: string
          enum: [rent, sale, both]
        status:
          type: string
          enum: [available, occupied, for_sale]
        description:
          type: ["string", "null"]
        photos:
          type: array
          items:
            type: string
        businessType:
          type: ["string", "null"]
          description: For shops/grocery stores
        amenities:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, type, sizeSqft, address, locality, purpose, status, photos, amenities, createdAt, updatedAt]
...
    Tenant:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        aadharNumber:
          type: ["string", "null"]
        emergencyContact:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        joinedAt:
          type: string
          format: date-time
        notes:
          type: ["string", "null"]
      required: [id, name, email, phone, status, joinedAt]
...
    Lease:
      type: object
      properties:
        id:
          type: integer
        propertyId:
          type: integer
        tenantId:
          type: integer
        property:
          oneOf:
            - $ref: "#/components/schemas/Property"
            - type: "null"
        tenant:
          oneOf:
            - $ref: "#/components/schemas/Tenant"
            - type: "null"
        startDate:
          type: string
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        rentAmount:
          type: number
        rentPeriod:
          type: string
          enum: [daily, weekly, monthly, yearly]
        securityDeposit:
          type: ["number", "null"]
        status:
          type: string
          enum: [active, expired, terminated]
        autoRenew:
          type: boolean
        lateFeePercentage:
          type: ["number", "null"]
        gracePeriodDays:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, propertyId, tenantId, startDate, rentAmount, rentPeriod, status, autoRenew, createdAt]
    CreateLeaseBody:
      type: object
      properties:
        propertyId:
          type: integer
        tenantId:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        rentAmount:
          type: number
        rentPeriod:
          type: string
          enum: [daily, weekly, monthly, yearly]
        securityDeposit:
          type: ["number", "null"]
        autoRenew:
          type: boolean
        lateFeePercentage:
          type: ["number", "null"]
        gracePeriodDays:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
      required: [propertyId, tenantId, startDate, rentAmount, rentPeriod, autoRenew]
...
    MaintenanceRequest:
      type: object
      properties:
        id:
          type: integer
        propertyId:
          type: integer
        tenantId:
          type: integer
        property:
          oneOf:
            - $ref: "#/components/schemas/Property"
            - type: "null"
        tenant:
          oneOf:
            - $ref: "#/components/schemas/Tenant"
            - type: "null"
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [plumbing, electrical, carpentry, cleaning, pest_control, other]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        resolvedAt:
          type: ["string", "null"]
          format: date-time
        adminNotes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, propertyId, tenantId, title, description, category, priority, status, createdAt]
    CreateMaintenanceBody:
      type: object
      properties:
        propertyId:
          type: integer
        tenantId:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [plumbing, electrical, carpentry, cleaning, pest_control, other]
        priority:
          type: string
          enum: [low, medium, high, urgent]
      required: [propertyId, tenantId, title, description, category, priority]
    UpdateMaintenanceBody:
      type: object
      properties:
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        adminNotes:
          type: ["string", "null"]
      required: [status]
...